+++ packages/components/organisms/f-checkout/src/store/tests/checkout.module.test.js
@@ -5,7 +5,7 @@ import basketDelivery from '../../demo/get-basket-delivery.json';
 import checkoutAvailableFulfilment from '../../demo/checkout-available-fulfilment.json';
 import customerAddresses from '../../demo/get-address.json';
 import geoLocationDetails from '../../demo/get-geo-location.json';
-import { defaultCheckoutState, mockAuthToken } from '../../components/_tests/helpers/setup';
+import { defaultCheckoutState, mockAuthToken, mockAuthTokenNoNumbers, mockAuthTokenNoMobileNumber } from '../../components/_tests/helpers/setup';
 import { version as applicationVerion } from '../../../package.json';
 import { VUEX_CHECKOUT_ANALYTICS_MODULE, DEFAULT_CHECKOUT_ISSUE } from '../../constants';
 
@@ -294,7 +294,6 @@ describe('CheckoutModule', () => {
     describe('actions ::', () => {
         let commit;
         let dispatch;
-
         let payload;
 
         beforeEach(() => {
@@ -312,6 +311,7 @@ describe('CheckoutModule', () => {
 
         describe('getCheckout ::', () => {
             let config;
+            let checkoutDeliveryCopy;
 
             beforeEach(() => {
                 config = {
@@ -322,7 +322,10 @@ describe('CheckoutModule', () => {
                     timeout: payload.timeout
                 };
 
-                axios.get = jest.fn(() => Promise.resolve({ data: checkoutDelivery }));
+                // Use a new copy per test so any mutations do not affect subsequent tests
+                checkoutDeliveryCopy = Object.assign(checkoutDelivery);
+
+                axios.get = jest.fn(() => Promise.resolve({ data: checkoutDeliveryCopy }));
             });
 
             it(`should get the checkout details from the backend and call ${UPDATE_STATE} mutation.`, async () => {
@@ -331,7 +334,7 @@ describe('CheckoutModule', () => {
 
                 // Assert
                 expect(axios.get).toHaveBeenCalledWith(payload.url, config);
-                expect(commit).toHaveBeenCalledWith(UPDATE_STATE, checkoutDelivery);
+                expect(commit).toHaveBeenCalledWith(UPDATE_STATE, checkoutDeliveryCopy);
             });
 
             it(`should call '${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill' mutation with an array of updated field names.`, async () => {
@@ -340,6 +343,87 @@ describe('CheckoutModule', () => {
 
                 // Assert
                 expect(dispatch).toHaveBeenCalledWith(`${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill`, state, { root: true });
+            });
+
+            describe('when the `customer` model is not returned from the api', () => {
+                it('should not error when checking phone number', async () => {
+                    // Arrange
+                    checkoutDeliveryCopy.customer = null;
+
+                    // Act
+                    await getCheckout({ commit, state, dispatch }, payload);
+
+                    // Assert
+                    expect(checkoutDeliveryCopy.customer).toBe(null);
+                    expect(axios.get).toHaveBeenCalledWith(payload.url, config);
+                    expect(commit).toHaveBeenCalledWith(UPDATE_STATE, checkoutDeliveryCopy);
+                });
+            });
+
+            describe('when the customers phone number is returned from the api', () => {
+                it('should not use neither of AuthToken phone numbers', async () => {
+                    // Arrange
+                    const expectedPhoneNumber = '5678901234';
+                    checkoutDeliveryCopy.customer = {
+                        phoneNumber: expectedPhoneNumber
+                    };
+
+                    // Act
+                    await getCheckout({ commit, state, dispatch }, payload);
+
+                    // Assert
+                    expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
+                });
+            });
+
+            describe('when the customers phone number is not returned from the api', () => {
+                beforeEach(() => {
+                    checkoutDeliveryCopy.customer = {
+                        phoneNumber: null
+                    };
+                });
+
+                afterEach(() => {
+                    // Reset the AuthToken for subsequent tests
+                    state.authToken = authToken;
+                });
+
+                it('should assign the AuthToken mobile number to the `customer.phoneNumber`', async () => {
+                    // Arrange
+                    state.authToken = mockAuthToken;
+                    const expectedPhoneNumber = '9876543210'; // AuthToken Mobile No.
+
+                    // Act
+                    await getCheckout({ commit, state, dispatch }, payload);
+
+                    // Assert
+                    expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
+                });
+
+                it('should assign the AuthToken phone number to the `customer.phoneNumber` if the AuthToken mobile number is missing', async () => {
+                    // Arrange
+                    state.authToken = mockAuthTokenNoMobileNumber;
+                    config.headers.Authorization = `Bearer ${state.authToken}`;
+                    const expectedPhoneNumber = '0123456789'; // AuthToken Phone No.
+
+                    // Act
+                    await getCheckout({ commit, state, dispatch }, payload);
+
+                    // Assert
+                    expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
+                });
+
+                it('should assign nothing to the `customer.phoneNumber` if both the AuthToken phone numbers are missing', async () => {
+                    // Arrange
+                    state.authToken = mockAuthTokenNoNumbers;
+                    config.headers.Authorization = `Bearer ${state.authToken}`;
+
+                    // Act
+                    await getCheckout({ commit, state, dispatch }, payload);
+
+                    // Assert
+                    expect(checkoutDeliveryCopy.customer.phoneNumber).toBeUndefined();
+                });
             });
         });