+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v3.29.0
+------------------------------
+*April 16, 2021*
+
+## Added
+- `husky` to enable pre-commit lint / test checks 
+
+
+v3.28.0
+------------------------------
+*April 14, 2021*
+
+## Added
+- Test tagging mechanism for chrome / browserstack
+
+
 v3.27.0
 ------------------------------
 *April 13, 2021*