+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -4,12 +4,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v0.92.0
+v0.93.0
 ------------------------------
 *April 19, 2021*
 
 ### Added
 - Handling of duplicate orders.
+
+### Changed
+- Refactored handling order issues.
+
+
+v0.92.0
+------------------------------
+*April 16, 2021*
+
+### Added
+- Use AuthToken phone nos. if checkout api fails to return a customer phone no.
+
 
 ### Changed
 - Refactored handling order issues.